Use RXY API instead of RX API to generate RXY instructions
As part of eclipse/omr#3113 we are making instruction kinds immutable.
As such we no longer support "upgrading" RX instructions to RXY at
instruction selection. This means that proper APIs must be used to
construct RXY instructions.

In this instance the mnemonic we are using is LY or LG, both of which
are RXY instructions. It is functionally incorrect to use the
`generateRXInstruction` API to generate such instructions.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>