Add missing HPR preserved registers to CHelper linkage on 64-bit
On 64-bit platforms when GPRx is marked as a preserved register we
should also mark HPRx as preserved as well. The linkage logic will look
at all the registers and any register not marked as preserved will be
assumed to be volatile and a dummy dependency will be created for it.

If GPRx is marked as preserved and HPRx is not then the linkage logic
will create a dummy dependency for HPRx. On 64-bit platforms all dummy
dependency virtual registers are treated as true 64-bit registers and
as such the dummy dependency on HPRx will force the register allocator
to evict any 64-bit value from GPRx, unnecessarily so (because GPRx is
actually preserved).

To avoid this performance issue we force HPRx to be preserved if GPRx
is also preserved (on 64-bit targets).

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>