Make SCQueryFunctions helpers inline
Defining static utility functions in header files is typically not
wanted because each compilation unit which includes the header will get
it's own copy of the defined functions. Moreover when building with
warnings as errors (which is the default today) we cannot compile such
compilation units without using all defined APIs as the compiler will
emit a(n) (warning) error.

We define the functions within SCQueryFunctions.h as inline functions
to avoid such problems.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>