Remove unneeded code in NULLCHKEvaluator
There is code which decrements and increments node counts for stores and
loads which is unnecessary. In the case of stores, the IL trees do not
produce a value so decrementing and then incrementing the node does
nothing at all. The comment seemingly implies that the result may return
NULL (presumably from the evaluators) however such a value is never
consumed so it should not matter.

In the case of loads we seem to be trying to avoid a evaluating a load
with exactly one unevaluated reference underneath a NULLCHK which will
never happen because such IL will have been purged from the method
during simplification passes at the optimizer level.

With all this, there is no need to replicate such logic across all the
platforms, and we simply delete the code.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>