Promote inlining of java/lang/reflect/Method.invoke
The `Method.invoke` API contains a call to `Reflect.getCallerClass()`
API which when executed will trigger a stack walking operation.
Performance wise this is quite expensive. The `Reflect.getCallerClass()`
API returns the class of the method which called `Method.invoke`, so if
we can promote inlining of `Method.invoke` we can eliminate the
`Reflect.getCallerClass()` call with a simple load, thus avoiding the
expensive stack walk.